+++ crates/pgt_workspace/src/workspace.rs
@@ -21,7 +21,7 @@ use crate::{
 mod client;
 mod server;
 
-pub(crate) use server::StatementId;
+pub(crate) use server::document::{Document, Statement, StatementId};
 
 #[derive(Debug, serde::Serialize, serde::Deserialize)]
 #[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
